[winnie.ka] 방탈출 인증 관리 1, 2단계 (#36)
* feat: 토큰 발급 api 구현

* feat: 내 정보 조회 api 구현

* feat: 인증 관련 처리 추가

* refactor: 테스트 리팩토링

* feat: 인증된 사용자만 예약을 삭제할 수 있도록 변경

* refactor: 자잘한 리팩토링

* fix: reservation을 생성할 때 멤버 정보는 username으로 생성

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -1,7 +1,9 @@
 package nextstep.reservation;
 
+import nextstep.member.Member;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
+import nextstep.support.AuthorizationException;
 import nextstep.support.DuplicateEntityException;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
@@ -34,7 +36,7 @@ public class ReservationService {
 
         Reservation newReservation = new Reservation(
                 schedule,
-                reservationRequest.getName()
+                reservationRequest.getUsername()
         );
 
         return reservationDao.save(newReservation);
@@ -49,10 +51,13 @@ public class ReservationService {
         return reservationDao.findAllByThemeIdAndDate(themeId, date);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, Member principal) {
         Reservation reservation = reservationDao.findById(id);
         if (reservation == null) {
             throw new NullPointerException();
+        }
+        if (!reservation.getUsername().equals(principal.getUsername())) {
+            throw new AuthorizationException();
         }
 
         reservationDao.deleteById(id);
